remove multiline, remove spell name from args
You don't need to use multiline when returning a single command using an alias, and I would avoid passing the spell name in the args when using `!cast` as it might activate user's snippets.  I provided alternatives to this but I didn't test it (I pulled the cArgs line from an alias of mine and removed multiline and \n before {ctx.prefix} )

So, best to test it first.

Signed-off-by: TheReverendB <75177105+TheReverendB@users.noreply.github.com>